+++ libs/components/src/lib/searchable-select/locale.ts
@@ -3,6 +3,10 @@ export interface SearchableSelectLocale {
 	noOptionsMessage: string;
 	noMatchesMessage: string;
 	loadingOptionsMessage: string;
+	selectAllLabel: string;
+	deselectAllLabel: string;
+	selectedAllMessage: string;
+	deselectedAllMessage: string;
 	removeTagButtonLabel: (label: string) => string;
 	optionSelectedMessage: (name: string) => string;
 	optionDeselectedMessage: (name: string) => string;