+++ libs/components/src/locales/de-DE.ts
@@ -176,6 +176,10 @@ const deDE: Locale = {
 		clearButtonLabel: 'Auswahl löschen',
 		noOptionsMessage: 'Keine Optionen',
 		noMatchesMessage: 'Keine Optionen gefunden',
+		selectAllLabel: 'Alles auswählen',
+		deselectAllLabel: 'Auswahl aufheben',
+		selectedAllMessage: 'Alle Optionen ausgewählt',
+		deselectedAllMessage: 'Alle Optionen abgewählt',
 		loadingOptionsMessage: 'Laden...',
 		removeTagButtonLabel: /* istanbul ignore next */ (label: string) =>
 			`${label} entfernen`,
@@ -250,8 +254,7 @@ const deDE: Locale = {
 				? `Weitere Informationen zu ${ariaLabel} anzeigen`
 				: `Weitere Informationen anzeigen`,
 	},
-	simpleColorPicker: {
-		colorPaletteLabel: 'Farbpalette',
+	baseColorPicker: {
 		colorSwatchLabel: /* istanbul ignore next */ (
 			value: string,
 			label?: string,
@@ -263,6 +266,19 @@ const deDE: Locale = {
 			if (selected) swatchLabel += `, ausgewählt.`;
 			return swatchLabel;
 		},
+	},
+	simpleColorPicker: {
+		colorPaletteLabel: 'Farbpalette',
+	},
+	colorPicker: {
+		popupLabel: 'Farbwähler',
+		swatchesLabel: 'Gespeicherte Farben:',
+		pickerButtonLabel: 'Farbwähler öffnen',
+		saveButtonLabel: 'Aktuelle Farbe speichern',
+		closeButtonLabel: 'Farbwähler schließen',
+		copyButtonLabel: 'Farbe in die Zwischenablage kopieren',
+		copyErrorText:
+			'Kopieren fehlgeschlagen. Zugriff auf die Zwischenablage wurde blockiert. Versuchen Sie es erneut.',
 	},
 };