+++ libs/test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [5.3.0](https://github.com/Vonage/vivid-3/compare/test-utils-v5.2.0...test-utils-v5.3.0) (2025-10-16)
+
+
+### Features
+
+* **option:** add `trailing-meta` slot (VIV-2792) ([#2524](https://github.com/Vonage/vivid-3/issues/2524)) ([6948f2b](https://github.com/Vonage/vivid-3/commit/6948f2be5e71e15f19fedf60c00bad98dea5ccb7))
+
 ## [5.2.0](https://github.com/Vonage/vivid-3/compare/test-utils-v5.1.0...test-utils-v5.2.0) (2025-09-26)